+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 3.1.0 / 2014-03-09
+
+* Added `restart`, `clean` and `status` commands.
+* The `address` configuration option is now `bind_addresses` and it supports multiple values.
+
 ### 3.0.8 / 2014-03-08
 
 * Installation fixes.